+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to
 - ✨(settings) Allow configuring PKCE for the SSO #886
 - 🌐(i18n) activate chinese and spanish languages #884
 - 🔧(backend) allow overwriting the data directory #893
+- ➕(backend) add  `django-lasuite` dependency #839
 
 ## Changed